+++ src/misc/init_game.c
@@ -3,13 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   init_game.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: shattori <shattori@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nando <nando@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/01 08:06:33 by shattori          #+#    #+#             */
-/*   Updated: 2025/10/16 16:35:54 by shattori         ###   ########.fr       */
+/*   Updated: 2025/10/17 17:38:56 by nando            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "cub3D.h"
 #include "game.h"
 #include "parse.h"
 
@@ -38,19 +39,25 @@ static char	*safe_strdup(const char *src)
 	return (dst);
 }
 
-// 🔹 2次元配列（マップ）をディープコピー
+static void	free_partial(char **map, int last)
+{
+	while (last-- > 0)
+		free(map[last]);
+	free(map);
+}
+
 static char	**duplicate_map(char **map)
 {
-	int		i;
-	int		count;
 	char	**new_map;
+	int		count;
+	int		i;
 
 	if (!map)
 		return (NULL);
 	count = 0;
 	while (map[count])
 		count++;
-	new_map = (char **)malloc(sizeof(char *) * (count + 1));
+	new_map = malloc(sizeof(char *) * (count + 1));
 	if (!new_map)
 		return (NULL);
 	i = 0;
@@ -58,12 +65,7 @@ static char	**duplicate_map(char **map)
 	{
 		new_map[i] = safe_strdup(map[i]);
 		if (!new_map[i])
-		{
-			while (i-- > 0)
-				free(new_map[i]);
-			free(new_map);
-			return (NULL);
-		}
+			return (free_partial(new_map, i), NULL);
 		i++;
 	}
 	new_map[count] = NULL;
@@ -88,41 +90,6 @@ t_data	*duplicate_data(t_data *src)
 	dst->player_y = src->player_y;
 	dst->player_dir = src->player_dir;
 	return (dst);
-}
-
-// 🔹 プレイヤー初期化（方角文字に応じて方向ベクトルとカメラ面を設定）
-void	init_player_from_data(t_player *p, t_data *data)
-{
-	p->pos_x = data->player_x + 0.5;
-	p->pos_y = data->player_y + 0.5;
-	if (data->player_dir == 'N')
-	{
-		p->dir_vec_x = 0;
-		p->dir_vec_y = DIR_BACKWARD;
-		p->plane_x = VIEWING_ANGLE;
-		p->plane_y = 0;
-	}
-	else if (data->player_dir == 'S')
-	{
-		p->dir_vec_x = 0;
-		p->dir_vec_y = DIR_FORWARD;
-		p->plane_x = -VIEWING_ANGLE;
-		p->plane_y = 0;
-	}
-	else if (data->player_dir == 'E')
-	{
-		p->dir_vec_x = DIR_RIGHT;
-		p->dir_vec_y = 0;
-		p->plane_x = 0;
-		p->plane_y = VIEWING_ANGLE;
-	}
-	else if (data->player_dir == 'W')
-	{
-		p->dir_vec_x = DIR_LEFT;
-		p->dir_vec_y = 0;
-		p->plane_x = 0;
-		p->plane_y = -VIEWING_ANGLE;
-	}
 }
 
 // 🔹 ゲーム全体の初期化
