+++ src/parse/check_texture_new.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_texture_new.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: shattori <shattori@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nando <nando@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 22:20:57 by shattori          #+#    #+#             */
-/*   Updated: 2025/10/17 16:39:14 by shattori         ###   ########.fr       */
+/*   Updated: 2025/10/17 17:45:47 by nando            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
