+++ src/parse/get_map.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_map.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: shattori <shattori@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nando <nando@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/17 00:13:01 by shattori          #+#    #+#             */
-/*   Updated: 2025/10/17 16:51:05 by shattori         ###   ########.fr       */
+/*   Updated: 2025/10/17 17:45:57 by nando            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
